Revert "tdf#164201 docx convert-to: prefer Word 2010–365 Document, not 2007"
This reverts 24.04 commit 858c452e5e6b7acad6932df7d1cd2f0950d47f33
on the 24.04 branch only. It is fine to remain on 25.04.

Reason for revert: There is a follow-up commit in bug 164201 that is necessary for Microsoft to be able to fully read the document. That necessary piece is only in 25.04. Making this seemingly simple change has a lot more implications than expected (primarily by Microsoft not being able to read files created according to their own specs). So I am not comfortable recommending that it be backported to our stable 24.04.

It was originally backported primarily to benefit mso-test - not a compelling reason to hang on to it.

The specific problem that lead to the follow-up commit was the loss of create/modified date/time fields (seen only in MS Word).

The specific problem that lead to this revert is content controls only showing the placeholder text, and not the real text (seen only in MS Word).

Change-Id: I2e3d0e835282938dff86d80b17b8bb22958a0622
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183223
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>