Related: tdf#128186 force key window to a native full screen window
AquaSalMenu::setDefaultMenu() is generally called when the key
window has been closed. When not in native full screen mode,
macOS appears to automatically set the key window.

However, closing a native full screen window sometimes causes
the application to drop out of full screen mode even if there
are still native full screen windows open. So, if the application
is active, activate all windows to force macOS to set the key
to a window rather than leaving the application in a state where
the key window is nil.

Also, add a EnableNativeFullScreenWindows expert preference so that
users can disable native full screen mode and make the green button
in the titlebar zoom instead of switching to a full screen window.

Change-Id: I3d9c538d784d7abcd5e75975ea00582768851ec7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183229
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>