cool#11357 sw redline reinstate: handle a single rich delete
In case the cursor is inside a delete redline (single-paragraph, plain
text), then .uno:ReinstateTrackedChange works for that change. If the
change contains formatting, then that formatting is not copied to the
insertion.

This is simply because SwEditShell::ReinstatePaM() gets the content of
the delete redline as an OUString. Doing a rich copy would preserve
formatting, but then
sw::DocumentContentOperationsManager::CopyImplImpl() would use
lcl_DeleteRedlines() to strip away content which is inside a delete
redline, so that won't work out of the box, either.

Fix the problem by copying the content of the delete redline to a
clipboard doc and inserting from there, and by introducing a new
bDeleteRedlines flag (defaults to true). When that flag is off, the copy
does what we need here: it preserves formatting, but doesn't strip away
content which was originally inside a delete redline.

Go via SwTransferable instead of invoking
sw::DocumentContentOperationsManager::CopyRange() directly, because this
way the inserted content will be inside an insert redline & we get
working undo/redo out of the box.

Change-Id: Id853844b7282c6b1c36cb0f4f8cc75d968e53dcd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183256
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>