ScriptForge prettyprint in APSO 1.4.4
The latest version of APSO (1.4.4) introduces
a new positional argument in the
  def console():
in apso_utile.py, named 'prettyprint'.

The def is used in
  SF_Exception.PythonShell()

The absence of the argument caused
an exception/regression.
The actual patch makes above method compatible
with latest and previous versions of APSO.

No impact on user scripts.
No impact on user documentation.

Change-Id: I0a90112d441bba550b330925742e5021665ecee6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183265
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins