nss: remove unneded patch
This is a follow-up to

commit acf84f9e21999327b831580691430db8c8fe8a71
Author: Xisco Fauli <xiscofauli@libreoffice.org>
Date:   Fri Mar 14 12:28:51 2025 +0100

    nss: remove obsolete target

In nss 3.103 or older, when OS_TARGET is not specified, it defaults to WIN95
anyway. See comment

-  When building on Windows, OS_TARGET now defaults to WIN95. To use the WINNT build
      configuration, specify OS_TARGET=WINNT.

in doc/rst/legacy/nss_releases/nss_3.15.4_release_notes/index.rst
thus, nspr is built with --enable-win32-target=WIN95 as the Makefile
does

ifeq ($(OS_TARGET),WIN95)
NSPR_CONFIGURE_OPTS += --enable-win32-target=WIN95
endif

consequently, the patch added in
commit 486c91147c587febdacf3fa95d56633dbd703d5f
Author: David Ostrovsky <david@ostrovsky.org>
Date:   Tue Feb 23 22:25:06 2016 +0100

    Bump nss to 3.22.1 and nspr to 4.12

can be removed. The reason why it was added initially
is unclear to me though

Change-Id: I9a2acc2b2cdd560fce9166c2e6e11c71ee21b5e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183282
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>