cool#11357 sw redline reinstate: handle deletes in selection
In case the cursor is inside a delete redline, then
.uno:ReinstateTrackedChange works for that change, but if the cursor is
a selection that covers one or more deletes, then this simply kills the
selection.

This happens because reinstating creates new redlines, but
SwEditShell::ReinstateRedlinesInSelection() iterates the redline table
while reinstating.

Fix the problem by creating a copy of the redline pointers and work on
that copy, this way all deletes are reinstated correctly.

The command has no explicit state yet, though.

Change-Id: Ia0e7b8365cb423ebf10d8f75aaff30d6495e1f7a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183291
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>