tdf#150623 reduce cost of ColorCfg
when switching tabs on this document, we end up in the ScChildrenShapes
performance sinkhole, which constructs and destructs ColorCfg a ton of
times.
Which should normally be free, but a bunch of logic has crept into the
outer class, and now we spend a lot of time unnecessarily adding and
removing an application event listener.

So push that logic down into ColorCfg_Impl, where it belongs, and now
tab switching goes from 2 seconds to under 1 second.

Change-Id: I4c7a508da0064a11bd4fa0b9251a91a6cda50c71
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183301
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins