tdf#165510 vcl: Fix incorrect adjustment of bidi MultiSalLayouts
The final step in laying out text with font substitutions is to drop
invalid glyphs from the base layout and position glyphs from
substitution layouts in their place. This change fixes a logic error in
this algorithm which was causing tofu and overlapping characters in the
frequent case of an all-LTR base layout and an all-RTL substitution
layout containing multiple fallback runs.

Change-Id: I0d8b03671a77c7d4c41c9ea0f32b526ed44477fe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183340
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>