related tdf#165521 svx: don't start layout until text is set
It has been this way since it svdoashp.cxx
was introduced in 2004.

Calculating the layout is expensive.
So why turn on the layout updates early,
when both SetFixedCellHeight (which can cause a FormatFullDoc)
as well as SetText (which always triggers a full layout)
could still obsolete any previous layout results.

Change-Id: I258ae8ffc33cf6221f56a4eb52f151e10d08bb82
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183314
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins