cool#11357 sw redline reinstate: add command state
Go outside a redline with your cursor, the accept and reject UNO
commands report disabled state, but reinstate is enabled all the time.

This happens because SwView::GetState() handles FN_REDLINE_ACCEPT_DIRECT
/ FN_REDLINE_REJECT_DIRECT explicitly.

Fix the problem by handling FN_REDLINE_REINSTATE_DIRECT the same way: if
reject is disabled, then reinstate ("reject with history") is also OK to
be disabled.

With this, reinstate is mostly usable, so add it to the Edit -> Track
changes menu.

Change-Id: Ic534bc314f0910b1a8376cabe944bfb6f0bc6953
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183338
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins