tdf#165020 sw DoNotCaptureDrawObjsOnPage: fix left padding for wrap through
Open the bugdoc, the left shapre is partially outside the page frame,
but is not captured. This is fine, but the text should be clipped in a
way that only the right half is visible, and instead the left half is
visible.

This went wrong in commit 61692b82ae4ff62662509b5979a7aabc7d380678
(tdf#161635 sw DoNotCaptureDrawObjsOnPage: fix handling of left padding,
2024-06-19), where I assumed that in case the left paddig is increased
in Word for these partially off-page shapes, then that happens
unconditionally.

Fix the new use-case by only doing this for non-wrap-through shapes:
this restores the lack of increased left padding for the new use-case
and keeps the old one fixed.

Special-casing the 'wrap mode == through' case for off-page positioning
is not new, SwAnchoredObject::IsDraggingOffPageAllowed() and
SwAnchoredObjectPosition::GetInfoAboutObj() also takes it into account.

Change-Id: I07dec72e860c810a753c2c356f8f32cc2178b012
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183339
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 44b4ff646ea21df75f9e35b885d76d1aa0d61682)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183343