tdf#88908 sw: Add BalanceSpacesAndIdeographicSpaces compat option
Added a new Writer compatibility option, as noted. When enabled, Writer
adjusts the width of certain normal space characters to one half the
width of a CJK ideographic space. This adjustment is only applied to
leading and trailing spaces, multiple sequential spaces, and spaces
between CJK characters.

This flag is enabled automatically when importing DOCX files with the
w:balanceSingleByteDoubleByteWidth option enabled.

This flag is also enabled automatically when importing DOC files with
the fDntBlnSbDbWid compatibility flag unset.

Change-Id: I1c666a9894f2e0f302df57203b1cf488cc13000b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183412
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>