Suppress -Werror,-Wunused-macros
...during a public.ecr.aws/allotropia/libo-builders/lode based ASan/UBSan build,

> sal/osl/all/log.cxx:22:9: error: macro is not used [-Werror,-Wunused-macros]
> #define _SILENCE_EXPERIMENTAL_FILESYSTEM_DEPRECATION_WARNING
>         ^

(where the compiler happens to be Clang, so use a `#pragma clang` for
simplicity, even if this has probably more to do with the version of libstdc++
being used than with the actual compiler).  This is in non-master code that was
added to resolve conflicts in cherry-pick
9fb2fa9bddef863c1e444861f939d4d7956dcc4b "Best effort to create directories of
SAL_LOG_FILE".

Change-Id: I0c394180567526d65d61afa5aa8afecff67633a3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183376
Tested-by: allotropia jenkins <jenkins@allotropia.de>
Reviewed-by: Stephan Bergmann <stephan.bergmann@allotropia.de>