cool#11357 sw redline reinstate: add a reinstate-and-next command
Accept and reject has a way to do their action & move to the next
redline, reinstate is missing this combo.

This is because these versions are separate UNO commands and reinstate
just has the direct .uno:ReinstateTrackedChange.

Fix the problem by adding a new .uno:ReinstateTrackedChangeToNext that
does this, together with the command state.

Also add it to the menubar.

Change-Id: Ica2358b3fc38cc66e801292ab13f8ba3ec4e9ece
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183410
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>