cool#11357 sw redline reinstate: add a reinstate-end-next command
Accept and reject has a way to do their action & move to the next
redline, resinstate is missing this combo.

This is because these versions are separate UNO commands and reinstate
just has the direct .uno:ReinstateTrackedChange.

Fix the problem by adding a new .uno:ReinstateTrackedChangeToNext that
does this, together with the command state.

Also add it to the menubar.

Change-Id: Ica2358b3fc38cc66e801292ab13f8ba3ec4e9ece
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183363
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>