tdf#165933 avoid w:delText inside w:moveFrom
w:delText should only be used inside a w:del, we differentiate if we use
w:del or w:moveFrom by looking if the current redline is move and if it
is inside a move bookmark. But sometimes the run contents are writen
before the surrounding bookmarks, in that case we wrongly used delText.
To avoid this we need to check not only the existing bookmarks but also
the ones that will be opened in this run.

Change-Id: I6653d18545cc5eb5f302c5753cdd7dfb24bee6fc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183458
Reviewed-by: Jaume Pujantell <jaume.pujantell@collabora.com>
Tested-by: Jenkins