Related: tdf#161623 disable menubar visibility if no key window
If a window is in LibreOffice's internal full screen mode
and not in native full screen mode and then the user switches
to a different application and back using the Command-Tab keys.
the menubar and Dock would unexpectedly appear.

It appears that the key window will still be nil in this
case, so only enable menubar visibility if the key window
is not nil.

Change-Id: I1ad9aefe8183908044042efc2ec67df3bae252dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183500
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins