Related: tdf#161623 don't set the menubar to null on macOS
When a window enters LibreOffice's internal full screen mode,
the vcl code will hide the macOS menubar. However, if the
window is also in native full screen mode, macOS will force
the menubar to be visible.

While the vcl code already partially handles this case by
disabling all menu items when in LibreOffice's internal full
screen mode, the problem is that any submenus that were not
displayed before setting the menubar to null will show all
menu items with no title.

A simple way to reproduce this bug is to open a new Writer
or Calc document and do the following:
- Switch the window to LibreOffice's internal full screen
  mode by manually selecting the View > Full Screen menu
  item (the bug does not occur if its key shortcut is
  pressed)
- Switch the window to native full screen mode
- Click on the menubar and note that many of the submenus
  are displayed with menu items, but none of the menu items
  have a title

So, we need to keep the menubar visible and rely on the vcl
code to disable all menu items.

Change-Id: I97fc63a8cd028c029a71899090185bc09d4aa492
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183508
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins