chart a11y: Simplify XAccessible{,Extended}Component implementations
Let the base class AccessibleBase implement
XAccessibleExtendedComponent with default implementations
for the methods that simply return empty strings.

This simplifies the implementation in subclasses, that
now only need to override XAccessibleComponent
and XAccessibleExtendedComponent methods when they
actually want to have a different behavior, not
to avoid ambiguity due to XAccessibleComponent
being a base class of both of their base classes
(XAccessibleExtendedComponent, AccessibleBase).

Change-Id: I2a6a39a721de19178954db898d52a2d7bc01fcb7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183523
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>