chart a11y: Use OAccessibleExtendedComponentHelper
Let chart's AccessibleBase derive from
OAccessibleExtendedComponentHelper to make use of the
logic already implemented there instead of having to
implement all the XAccessible(Extended)Component
and XAccessibleEventBroadcaster methods manually.

AccessibleBase::getBounds implements what is needed
to implement OAccessibleExtendedComponentHelper::implGetBounds,
so rename the method.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the base class.

Adjust subclasses as well to deduplicate logic there, too.

No change in behavior intended or observed initially
when testing this with Accerciser and Orca and the gtk3 and qt6
VCL plugins for a chart inserted in Writer ("Insert" -> "Chart").

Change-Id: I5d29c099dc31d70711f4017e52a0859f22c5aa23
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183524
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>