tdf#165742 Step 4.1: Establish a narrow export path for chartex
This is a subtask of tdf#165742: Chartex charts are lost on input from OOXML
and re-export.

Build out the some of the export structure for chartex charts. This now
produces an output chartEx*.xml file with no c-namespace content, for
funnel charts only, and what appears to be fully well-formed OOXML
per the office/drawing/2014/chartex schema. However, MSO doesn't load
the result successfully, so I'm calling it step 4.1.

For the bulk of the output code in chartexport.cxx, there was a design
decision whether to (1) have a single 'if (chartex) ... else ...'
branch point high up, or to (2) branch much more granularly within the
code. (1) has the benefit of cleaner reading, without so many ifs, at
the expense of a lot of duplicated code. (2) allows for much less code
overall but with lots of ifs. For the time being I've taken approach
(2), not necessarily because it's the right one, but mostly because it
is much easier to move from (2) to (1) (largely by just duplicating
the code and in each large branch deleting the local branches that
don't apply), than from (1) to (2) (which would require examining the
differences between the two large branches in detail, in order to re-
integrate the code).

Change-Id: I78de1484359c358b99320b417ab68cd90dab3e55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183561
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit dd8c544e6128e5d3454c006e00f8f48e5f59a1e0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185263
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>