cool#11357 sw redline reinstate: fix undo string for a single redline
Open a document with an insert redline, reinstate, check the undo stack,
it says we have a "delete" on it.

This is technically true, but the user did a reinstate, which happens to
be implemented as a delete for inserts.

Fix the problem by adding a dedicated SwUndoId::REINSTATE_REDLINE and
use that when reinstating a single redline.

The "reinstate multiple redlines at once" case still needs fixing.

Change-Id: Ib2c58ac41fcb86b2f947b041ada318ea2a361e8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183573
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>