tdf#130857 qt weld: Select whole tree view row
weld::TreeView::select only takes a row number.
Select the whole row, not just the first cell
in that row.

This e.g. makes a difference when using the
"Move Up" button in the "Form" -> "Activation Order"
dialog once support for that dialog is declared
(which will happen in an upcoming commit).

Change-Id: Ib5d2ce5de02f2d4e552e28163ed978176be466ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183585
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins