cool#11497 lok: calc: update separators per locale
On LOK, multiple locales might be in use at the same time, so skipping
the update of native symbols and separators on load is not viable.

I've also tested the original bug for which this code is skipped and it
seem to not affect online in the same way.

Change-Id: Ic2e6e329c9bdfb55f6a5fdd540d102c05ac168cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183594
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit d6135b62ee9716f120e6ba347503cc222e364829)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183630
Tested-by: Jenkins
Reviewed-by: Jaume Pujantell <jaume.pujantell@collabora.com>