cool#11357 sw redline reinstate: fix undo count & string for multiple redlines
Open a document with insert redlines, reinstate, check the undo stack,
it says it has multiple undo actions.

This is technically true, but the user did a reinstate, which happens to
be implemented as deletions.

Fix the problem by using SwUndoId::REINSTATE_REDLINE in
SwEditShell::ReinstateRedlinesInSelection(), too.

This fixes both the undo count (now one reinstate results in one
user-visible action) & the string to say reinstate.

Change-Id: I8f59fa6334876ef1bd860b7aa41c0ed7d35cbf31
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183607
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>