sw: Resave notebookbar.ui with glade 3.40
... to minimize the unrelated diff whenever the file gets
edited in glade next time.

However, for now manually undo misc changes like automatically added

    <property name="can-focus">False</property>

for GtkToolButton objects as I'm not sure whether those might
have any side-effect (so might be best to handle separately
instead of having it in a change mostly changing underscore
to hyphen).

Suppress critical gla11y warnings like

    sw/uiconfig/swriter/ui/notebookbar.ui: FATAL WARNING: 'GtkMenuItem' 'Menu-Extension-PackageManagerDialog' does not have its own label

as false positives, because labels for Notebookbar items are
set in C++ code, not the .ui file.

This .ui file is used for the Notebookbar in Writer
when "View" -> "User Interface" is set to "Tabbed".

Change-Id: I543b692f1d26404b0def59dfb8b203c7cbb1dbe1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183609
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>