editeng a11y: Rely on XAccessibleContext implementing XAccessiblecomponent
Only query the XAccessibleContext of the parent
accessible for the XAccessibleComponent interface,
not the XAccessible first and the XAccessibleContext
only as fallback.

As far as I understand it, the XAccessible is only there
to allow retrieving the XAccessibleContext using
XAccessible::getAccessibleContext and all other logic
should be implemented in the class implementing
the XAccessibleContext interface.
That's also what the platform bridges rely on, see
e.g. `getComponent` in vcl/unx/gtk3/a11y/atkcomponent.cxx.

If there's any case where this is handled differently,
that one should likely be adjusted to match the pattern
used everywhere else, too, instead of having special
handling here in AccessibleEditableTextPara.

(In most cases, it's the same class that implements both,
the XAccessible and the XAccessibleContext interface
anyway.)

Change-Id: Ic03c11f8406213f5f44074e39fe867b4e8b93dff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183624
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>