editeng a11y: Drop paragraph's a11y description
So far, "paragraph <number>" and the (start of) the
paragraph text were reported as the accessible description
of the paragraph.

This resulted e.g. in Orca saying something like
"Paragraph 0: Hello world Hello world" when moving cursor
to a paragraph containing the text "Hello world" in Impress,
because both, a11y description and the text reported via the
AtspiText interface were reported.

Drop the a11y description, which is also consistent with Writer
where no description is reported. The

    // (writer takes first sentence here, but that's not supported
    // from EditEngine)

comment dropped along with this commit looks outdated.

The fact that this is a paragraph is already reported via the
a11y role and the text is reported via the XAccessibleText
interface. (The information which paragraph this is could
be retrieved by checking the children of the paragraph's
parent, so if an AT wanted to report that, it could
implement corresponding logic on the client side.)

Adjust the Impress a11y test originally added in

    commit 6185a27db46bf5cba404e669eaf3a1c78f4a8607
    Author: Colomban Wendling <cwendling@hypra.fr>
    Date:   Sun Jul 31 17:20:03 2022 +0200

        tdf#150064 Add tests for a11y tree order

        Check the tree order is stable across loads and page switches.

accordingly, which included the description in
the expected output, but that's not the purpose
of the test.

Also adjust the sw a11y test
(test::AccessibleTestBase::BasicTestFontworkDialog) accordingly.

Change-Id: Ifbc1a15a976d7488396630c6bc3c9e42ba26ccc1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183625
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>