editeng a11y: Drop AccessibleStaticTextBase::mpThis etc.
AccessibleStaticTextBase::mpThis is only used to
pass it to Exception ctors, so it seems a bit
overkill to have to manage it for that purpose.

Drop it and the corresponding setter
AccessibleStaticTextBase::SetEventSource.
Its doc says that the interface is also used
as the event source for a11y events, but that's
not actually the case.

Also drop ScAccessibleCell::Init which now no
longer needs to override the base class implementation,
ScAccessibleCellBase::Init.

Change-Id: Ide64cf2e59bfb9ae7df41171cc36e1df0c86eb9f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183628
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>