sd a11y test: Add a shape to layout test
Extend the existing layout test by adding a shape object
into the document.

The shape has a name of "MyRectangle" and a description/Alt text
of "Some Rectangle" set and contains the text "Text inside Shape".

Relevant diff in the fodp file for the new shape:

    +    <draw:custom-shape draw:name="MyRectangle" draw:style-name="gr4" draw:text-style-name="P11" draw:layer="layout" svg:width="6cm" svg:height="4cm" svg:x="6cm" svg:y="10.5cm">
    +     <svg:desc>Some Rectangle</svg:desc>
    +     <text:p text:style-name="P10">Text inside Shape</text:p>
    +     <draw:enhanced-geometry svg:viewBox="0 0 21600 21600" draw:type="rectangle" draw:enhanced-path="M 0 0 L 21600 0 21600 21600 0 21600 0 0 Z N"/>
    +    </draw:custom-shape>

(The rest is mostly noise from opening + resaving the existing
document.)

The main reason to add this is to have something to
test the accessible description.
Currently, the expected test output includes a test
for the a11y description of various shapes being
" " (i.e. a single space character, see the implementation
in AccessibleShape::getAccessibleDescription that falls
back to this), but that will be changed to an empty string in
an upcoming commit.

The newly added shape will keep having a non-empty
description, so that logic can still be tested.

Change-Id: I92a781aaf9c05df6264e64a59db775495fa5671c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183674
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins