svx a11y: Don't fall back to space character for shape a11y description
Don't use " " (i.e. a string containing a single
space character) as the fallback for the accessible
description of a shape, but use an empty string instead.

Adjust the existing Impress a11y test accordingly, which
included the description reported so far in the
expected output, but that's not the purpose of the
test.

The text was previously extended in

    Change-Id: I92a781aaf9c05df6264e64a59db775495fa5671c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Apr 2 16:53:58 2025 +0200

        sd a11y test: Add a shape to layout test

to ensure the logic for testing a11y descriptions
keeps being used even with this new commit in place.

Also adjust a CppunitTest_sw_a11y dialog test accordingly.

Change-Id: If4bd88cf5e2dd48c68fa93a69871d0bcfbee9424
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183675
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>