tdf#60700 stop creating empty Configurations2/images in ODFs
This change stops the creation of empty Configurations2/images dirs
on new documents that do not have anything that needs to be stored
there. Previously, LibreOffice would always create this directory.

Whenever an XStorage is opened readwrite on a path, it will immediately
create that path. LibreOffice was opening all of these paths readwrite.
So, the solution is to always open these paths as readonly, and then
when we need to write data to them we close them and re-open them as
readwrite. If we never need to write data to a path, then it won't be
created.

This patch specifically stops the creation of subdirectories that were
being created by ImageManager.

Other patches in the relation chain handle paths created by
UIConfigurationManager and DocumentAcceleratorConfiguration.

Test coverage was added in b9054fed3724a19479e221a8b4818a127a18aabe

$ CPPUNIT_TEST_NAME="Test testTdf60700_images" make CppunitTest_sw_odfexport

I also manually tested with the following procedure:

1. Open a new document in writer.
2. Tools > Customize > Toolbars.
3. Set scope to "Untitled 1".
4. Add a new toolbar "New Toolbar 1".
5. Add any command from Available Commands to the new toolbar.
6. Add any icon
7. OK.
8. Visually verify that the toolbar is present with the custom icon.
9. Save the document.
10. Verify that the custom icon is saved correctly inside the file.

$ unzip -l Untitled\ 1.odt

Archive:  Untitled 1.odt
  Length      Date    Time    Name
---------  ---------- -----   ----
       39  2025-05-17 16:01   mimetype
      353  2025-05-17 16:01   Configurations2/toolbar/custom_toolbar_cba48a4e.xml
      444  2025-05-17 16:01   Configurations2/images/lc_imagelist.xml
      469  2025-05-17 16:01   Configurations2/images/Bitmaps/lc_userimages.png
        0  2025-05-17 16:01   Configurations2/accelerator/
    13358  2025-05-17 16:01   styles.xml
      899  2025-05-17 16:01   manifest.rdf
     3771  2025-05-17 16:01   content.xml
     1011  2025-05-17 16:01   meta.xml
    14746  2025-05-17 16:01   settings.xml
      318  2025-05-17 16:01   Thumbnails/thumbnail.png
     1406  2025-05-17 16:01   META-INF/manifest.xml
---------                     -------
    36814                     12 files

10. Close and reopen the document.
11. Visually verify that the toolbar with the custom icon is present.

Change-Id: I48b2b8f540f7640ae25da6f27065e7b07f955c7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185450
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>